// src/api/models/CatalogoDtoApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CatalogoDto } from './CatalogoDto';
export type CatalogoDtoApiResponse = {
    success?: boolean;
    data?: CatalogoDto;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};