// src/api/models/CatalogoDtoIEnumerableApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CatalogoDto } from './CatalogoDto';
export type CatalogoDtoIEnumerableApiResponse = {
    success?: boolean;
    data?: Array<CatalogoDto> | null;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};